Move deriveLiteral internal function to derive.go
