+++ pkg/ssg/vault/cipher.go
@@ -3,7 +3,6 @@ package vault
 import (
 	"crypto/aes"
 	"crypto/cipher"
-	"encoding/hex"
 	"fmt"
 	"io"
 	"strings"
@@ -21,34 +20,6 @@ func parse(in string) (string, string) {
 		return "", ""
 	}
 	return l[0], l[1]
-}
-
-func deriveLiteral(v Vault, keyp string, keysz int, ivp string, ivsz int) ([]byte, []byte, error) {
-	encoded, err := v.Provider.Get(keyp)
-	if err != nil {
-		return nil, nil, err
-	}
-	key, err := hex.DecodeString(string(encoded))
-	if err != nil {
-		return nil, nil, err
-	}
-
-	encoded, err = v.Provider.Get(ivp)
-	if err != nil {
-		return nil, nil, err
-	}
-	iv, err := hex.DecodeString(string(encoded))
-	if err != nil {
-		return nil, nil, err
-	}
-
-	if len(key) != keysz {
-		return nil, nil, fmt.Errorf("insufficient key size (%d bytes): want exactly %d bytes", len(key), keysz)
-	}
-	if len(iv) != ivsz {
-		return nil, nil, fmt.Errorf("insufficient initialization vector size (%d bytes): want exactly %d bytes", len(key), ivsz)
-	}
-	return key, iv, nil
 }
 
 func (c Cipher) stream() (cipher.Stream, cipher.Stream, error) {
